Add support for updating events
When an event is updated, its occurrences are deleted. Also, orphaned occurrences are cleaned up; this could be done elsewhere, perhaps on a schedule during small traffic hours. In any case, a change of primary key should be very rare, and so should be orphaned occurrences.

Also, after the `updated` event, the `saved` event is fired. This already creates occurrences, leading to their re-generation on update.